refactor indexer
